requirements.txt
absl-py==0.14.1
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
clang==5.0
colorama==0.4.4
cryptography==35.0.0
debugpy==1.4.3
decorator==5.1.0
distlib==0.3.1
Django==3.2.8
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
install==1.3.4
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.5
jupyter-core==4.8.1
Markdown==3.3.4
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.5.post0
prompt-toolkit==3.0.20
protobuf==3.18.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.15.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tornado==6.1
traitlets==5.1.0
twilio==7.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==20.4.4
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.1.0
